Changed the monitor.wds script to only flag on access violations and to output some crash details gave the fuzzer the ability to sort crash details in the same way that the triage tool does so that crashes can be quickly looked at without running the triage tool.  The triage tool should still be run to verify that all crashes occur upon second test.

+++ fuzzer.py
@@ -8,44 +8,19 @@
 # 		- WinDbg (http://msdn.microsoft.com/en-us/windows/hardware/gg463009.aspx)
 # 		- GFlags (http://msdn.microsoft.com/en-us/windows/hardware/gg463009.aspx)
 # 		- Radamsa (http://code.google.com/p/ouspg/wiki/Radamsa)
-# 		- window_killer python library
+# 		- window_killer (https://github.com/jlday/window_killer)
+# 		- triage (https://github.com/jlday/triage)
 ####################################################################################
 # Additional Installation Notes:
 # 		The following WinDbg Script needs to be placed in the same directory as
 # the WinDbg executable.  It should be called "monitor.wds".
 '''
 * monitor.wds - run this script upon launching an executable, 
-*      when a crash occurs a file named "crashDetector.txt" 
-*      is written to the local directory.  There should be 
-*      no contents in this file, the presence only indicates
-*      that a crash was found.
+*      when a crash occurs a file named "crash_details.txt" 
+*      is written with the output of !exploitable
 
 sxr
-sxd -c ".logopen crashDetector.txt; .logclose; q" -h av
-sxi asrt
-sxi aph
-sxi bpe
-sxi eh
-sxi clr 
-sxi clrn
-sxi cce
-sxd -c ".logopen crashDetector.txt; .logclose; q" -h dm
-sxd -c ".logopen crashDetector.txt; .logclose; q" -h gp
-sxd -c ".logopen crashDetector.txt; .logclose; q" -h ii
-sxd -c ".logopen crashDetector.txt; .logclose; q" -h ip
-sxd -c ".logopen crashDetector.txt; .logclose; q" -h dz
-sxd -c ".logopen crashDetector.txt; .logclose; q" -h iov
-sxd -c ".logopen crashDetector.txt; .logclose; q" -h ch
-sxd -c ".logopen crashDetector.txt; .logclose; q" -h isc
-sxi 3c
-sxi svh
-sxi sse
-sxd -c ".logopen crashDetector.txt; .logclose; q" -h sbo
-sxd -c ".logopen crashDetector.txt; .logclose; q" -h sov
-sxi vs
-sxd -c ".logopen crashDetector.txt; .logclose; q" -h wkd
-sxi wob
-sxi wos
+sxd -c "!load msec.dll; .logopen crash_details.txt; .echo ********************************************************************************; !exploitable; .echo ********************************************************************************; r; .echo ********************************************************************************; u; .echo ********************************************************************************; k; .echo ********************************************************************************; q" -h av
 
 '''
 ####################################################################################
@@ -90,6 +65,10 @@ options:
 -c [cpu sample time]
 	time delay used when calculating the CPU usage percentage.
 	Default: 0.5
+-d 
+	Save crash details along with the crash files.  Normally these can
+	be gained by triaging the crashes, but can be useful for browsing
+	results before triaging.  Only used when not in JIT Debugger Mode
 -r 
 	Enable mutations using the Radamsa test case generator.
 -j 
@@ -105,10 +84,13 @@ options:
 	for each instance of the target application that is spawned and only 
 	deals with windows belonging to the target's PID
 -e
-	Uses the window_killer to attempt to close the main window of the target
-	application before killing the process.  This is used to trigger any 
-	events that might occur upon closing the application normally for process
-	cleanup
+	Uses the window_killer to attempt to close the main window of the 
+	target application before killing the process.  This is used to 
+	trigger any events that might occur upon closing the application 
+	normally for process cleanup
+-z 
+	Attempt to unzip files and fuzz contents as well as fuzzing the files 
+	themselves 
 -v
 	Verbose Mode, includes progress updates and error messages
 -h 	
@@ -116,7 +98,7 @@ options:
 '''
 ####################################################################################
 # Imports:
-import subprocess, os, time, random, sys, getopt, shutil
+import subprocess, os, time, random, sys, getopt, shutil, zipfile
 import psutil
 ####################################################################################
 # Global Variables:
@@ -126,18 +108,20 @@ outputDir = "Crashes"
 WinDbgPath = "WinDbg"
 TestDir = "Tests"
 target_args = ""
-crashTxt = "crashDetector.txt"
+crashTxt = "crash_details.txt"
 radamsaPath = "radamsa-0.3.exe"
 reportEvery = 1000
 cpu_usage_sample = 0.5
 max_time = 10
 MutationRate = 0.05
 target = None
+save_crash_details = False
 radamsa = False
 jitDebugging = False
 useGflags = True
 kill_windows = False
 close_main = False
+fuzz_zipped = False
 verbose = False
 ####################################################################################
 # Functions:
@@ -229,11 +213,13 @@ def InitBaseFiles(dir=None):
 # is found
 def RunTest(file):
 	global verbose
+	global TestDir
 	global outputDir
 	global target_args
 	global jitDebugging
 	global target
 	global max_time
+	global save_crash_details
 	global kill_windows
 	global close_main
 	global cpu_usage_sample
@@ -251,7 +237,7 @@ def RunTest(file):
 		
 		if os.path.exists(crashTxt):
 			os.remove(crashTxt)
-		
+			
 		if not os.path.exists(TestDir):
 			os.mkdir(TestDir)
 			
@@ -276,8 +262,6 @@ def RunTest(file):
 					try:
 						if close_main: 
 							window_killer.CloseMain(proc.pid)
-						else:
-							proc.kill()
 					except:
 						pass
 					time.sleep(1)
@@ -332,8 +316,6 @@ def RunTest(file):
 				if proc != None and proc.status != psutil.STATUS_DEAD:
 					if close_main:
 						window_killer.CloseMain(proc.pid)
-					else:
-						proc.kill()
 				if debugger != None and debugger.status != psutil.STATUS_DEAD:
 					debugger.kill()
 			except KeyboardInterrupt:
@@ -355,7 +337,14 @@ def RunTest(file):
 				if verbose:
 					print "Crash Detected!!"
 					print "Saving crash file... " + file
-				os.remove(crashTxt)
+				if save_crash_details:
+					details = outputFile + "-details.txt"
+					shutil.move(crashTxt, details)
+					import triage
+					triage.outputDir = outputDir
+					triage.ProcessDetailsFile(details, outputFile)
+				else:
+					os.remove(crashTxt)
 			else:
 				os.remove(file)
 				os.remove(outputFile)
@@ -377,6 +366,13 @@ def RunTest(file):
 			pass
 	if windowKiller != None:
 		windowKiller.start_halt()
+
+# Generates a unique name to be used as an output file name + path for newly fuzzed files		
+def GenerateTestFileName(basename):
+	testFile = outputDir + os.sep + basename[basename.rfind("\\") + 1:basename.rfind(".")] + ("-0x%0.8X" % random.randint(0, 0xFFFFFFFF)) + basename[basename.rfind("."):]
+	while testFile == None or os.path.exists(testFile):
+		testFile = outputDir + os.sep + basename[basename.rfind("\\") + 1:basename.rfind(".")] + ("-0x%0.8X" % random.randint(0, 0xFFFFFFFF)) + basename[basename.rfind("."):]
+	return testFile
 		
 # The main fuzzer loop
 # Responsable for enabling and disabling GFlags and WinDbg as JIT debugger
@@ -388,6 +384,7 @@ def RunFuzzer():
 	global useGflags
 	global radamsaPath
 	global radamsa
+	global fuzz_zipped
 	global verbose
 	
 	testFile = None
@@ -405,24 +402,57 @@ def RunFuzzer():
 		if not os.path.exists(outputDir):
 			os.mkdir(outputDir)
 		
+		if not os.path.exists(TestDir):
+			os.mkdir(TestDir)
+		
 		while True:
 			if verbose and ((reportEvery > 1 and count % reportEvery == 1) or (reportEvery == 1 and count % reportEvery == 0)):
 				print "Working on file #" + str(count) 
 			file = PickFile()
-			testFile = outputDir + os.sep + file[file.rfind("\\") + 1:file.rfind(".")] + ("-0x%0.8X" % random.randint(0, 0xFFFFFFFF)) + file[file.rfind("."):]
-			while testFile == None or os.path.exists(testFile):
-				testFile = outputDir + os.sep + file[file.rfind("\\") + 1:file.rfind(".")] + ("-0x%0.8X" % random.randint(0, 0xFFFFFFFF)) + file[file.rfind("."):]
+			testFile = GenerateTestFileName(file)
 			if radamsa:
 				subprocess.call(radamsaPath + " -o " + testFile + " " + file)
 			else:
 				open(testFile, "wb").write(mutate(open(file, "rb").read()))
 			RunTest(testFile)
 			count += 1
+			
+			# Process zipped files
+			if fuzz_zipped and zipfile.is_zipfile(file):
+				archive = zipfile.ZipFile(file, "r")
+				for subfile in archive.namelist():
+					testFile = GenerateTestFileName(file)
+					if verbose and ((reportEvery > 1 and count % reportEvery == 1) or (reportEvery == 1 and count % reportEvery == 0)):
+						print "Working on file #" + str(count)
+					tempFile = TestDir + os.sep + "temp.tmp"
+					if radamsa:
+						open(tempFile, "wb").write(archive.read(subfile))
+						tempFileOut = tempFile[:tempFile.rfind('.')] + "2" + tempFile[tempFile.rfind('.') + 1:]
+						subprocess.call(radamsaPath + " -o " + tempFileOut + " " + tempFile)
+						shutil.move(tempFileOut, tempFile)
+					else:
+						open(tempFile, "wb").write(mutate(archive.read(subfile)))
+					
+					test = zipfile.ZipFile(testFile, "w")
+					for item in archive.namelist():
+						if item != subfile:
+							buffer = archive.read(item)
+							test.writestr(item, buffer)
+						else:
+							test.write(tempFile, subfile, zipfile.ZIP_DEFLATED)
+					test.close()
+					os.remove(tempFile)
+					RunTest(testFile)
+					count += 1
+				archive.close()	
 	except KeyboardInterrupt:
 		if useGflags:
 			DisableGFlags()
 		if testFile != None and os.path.exists(testFile):
 			os.remove(testFile)
+		if save_crash_details:
+			import triage
+			triage.CleanupFiles(outputDir)
 		raise KeyboardInterrupt()	
 		
 # Prints the command line usage if run as stand alone application.
@@ -443,18 +473,20 @@ def main(args):
 	global max_time 
 	global MutationRate 
 	global target
+	global save_crash_details
 	global radamsa
 	global jitDebugging 
 	global useGflags 
 	global kill_windows
 	global close_main
+	global fuzz_zipped
 	global verbose 
 	
 	if len(args) < 2:
 		PrintUsage()
 		exit()
 	
-	optlist, argv = getopt.getopt(args[1:], 'b:o:w:p:t:a:i:m:s:c:rjgkevh')
+	optlist, argv = getopt.getopt(args[1:], 'b:o:w:p:t:a:i:m:s:c:drjgkezvh')
 	for opt in optlist:
 		if opt[0] == '-b':
 			baseDir = opt[1]
@@ -476,6 +508,8 @@ def main(args):
 			MutationRate = float(opt[1])
 		elif opt[0] == '-c':
 			cpu_usage_sample = float(opt[1])
+		elif opt[0] == '-d':
+			save_crash_details = True
 		elif opt[0] == '-r':
 			radamsa = True
 		elif opt[0] == '-j':
@@ -486,6 +520,8 @@ def main(args):
 			kill_windows = True
 		elif opt[0] == '-e':
 			close_main = True
+		elif opt[0] == '-z':
+			fuzz_zipped = True
 		elif opt[0] == '-v':
 			verbose = True
 		elif opt[0] == '-h':
